Desktop/Python/requirements.txt
asgiref
astroid
backports.entry-points-selectable
cffi
colorama
cycler
distlib
Django
filelock
fonttools
image
isort
kiwisolver
lazy-object-proxy
matplotlib
mccabe
numpy
opencv-python
packaging
Pillow
platformdirs
pycparser
pydub
pylint
pyparsing
python-dateutil
scipy
six
sounddevice
sqlparse
termcolor
toml
tzdata
urllib3
virtualenv
Wave
wavio
wrapt
